fernly: Remove GPL-2 files, move to libgcc
We used to use source files taken from U-Boot to allow for division and
modular arithmetic.  With the move to a BSD license, this is no longer
acceptible.

As a workaround, use a libgcc taken from a standard gcc distribution.
Add the .a file, which has not been modified in any way.

Signed-off-by: Sean Cross <xobs@kosagi.com>